pyproject.toml
[project]
name = "py-drf-react"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "djangorestframework==3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.15",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "semantic-kernel>=1.36.0",
    "tenacity>=9.1.4",
]
